+++ server/TWS Foundation/TWS Business/Depots/LocationsDepot.cs
@@ -1,7 +1,6 @@
 ﻿using CSM_Foundation.Database.Entity;
 
 using TWS_Business.Entities;
-using TWS_Business.Entities.Employees;
 
 namespace TWS_Business.Depots;
 
@@ -24,4 +23,7 @@ public class LocationsDepot
     ///     Generates a new depot handler for <see cref="Location"/>.
     /// </summary>
     public LocationsDepot(Database Databases, IDisposer? Disposer = null) : base(Databases, Disposer) { }
+
+    public LocationsDepot() : base(new(), null) {
+    }
 }