+++ server/TWS Foundation/TWS Customer.Quality/Q_Features/Q_Bussines/Q_EmployeesService.cs
@@ -1,0 +1,116 @@
+﻿using CSM_Foundation.Core.Utils;
+using CSM_Foundation.Customer.Quality;
+using CSM_Foundation.Database.Entity;
+using CSM_Foundation.Database.Entity.Models;
+using CSM_Foundation.Database.Entity.Models.Input;
+using CSM_Foundation.Database.Entity.Models.Output;
+
+using TWS_Business.Depots;
+using TWS_Business.Entities;
+using TWS_Business.Entities.Employees;
+
+using TWS_Customer.Features.Business;
+using TWS_Customer.Quality.Factories;
+
+
+namespace TWS_Customer.Quality.Q_Features.Q_Bussines;
+
+public class Q_EmployeesService
+    : BQ_Service<IEmployeesService> {
+    public Q_EmployeesService()
+        : base(
+                [
+                    DatabaseFactories.BusinessDatabaseFactory,
+                ]
+            ) {
+
+    }
+
+    #region [BQ_Service] implementations
+    protected override IEmployeesService ServiceFactory() {
+        TWS_Business.Database BussinesDatabase = DatabaseFactories.BusinessDatabaseFactory();
+
+        IEmployeesDepot EmployeesDepot = new EmployeesDepot(BussinesDatabase, Disposer);
+
+        return new EmployeesService(EmployeesDepot);
+    }
+    #endregion
+
+    #region Private Methods/Functions
+    Employee GenerateMock(string Entropy) {
+        DateOnly date = new(2030, 11, 11);
+
+        Status status = Store(
+                new Status {
+                    Name = Entropy,
+                    Description = Entropy,
+                }
+            );
+        Status statusI = Store(
+                new Status {
+                    Name = "I" + Entropy,
+                    Description = Entropy,
+                }
+            );
+
+        Identification identification = Store(
+                 new Identification {
+                     Name = Entropy,
+                     Lastname = Entropy,
+                     Status = statusI,
+                 }
+            );
+
+        Employee_Dates employee_Dates = Store(
+        new Employee_Dates {
+            CNAP = date,
+            IMSS = date,
+            Hire = date,
+            Termination = date,
+        }
+            );
+
+        return new Employee {
+            CURP = Entropy + Entropy[..2],
+            RFC = Entropy[..13],
+            NSS = Entropy[..11],
+            Status = status,
+            Identification = identification,
+            Dates = employee_Dates,
+        };
+    }
+    #endregion
+
+    [Fact(DisplayName = "[View]: Records view")]
+    public async Task View() {
+        Store(GenerateMock(RandomUtils.String(16)));
+        SetViewOutput<Employee> viewOutput = await _service.View(
+                new OperationInput<Employee, SetViewInput<Employee>> {
+                    Parameters = new() {
+                        Retroactive = false,
+                        Range = 10,
+                        Page = 1,
+                    }
+                }
+            );
+
+        Assert.Multiple(
+            () => Assert.True(viewOutput.Pages > 0),
+            () => Assert.True(viewOutput.Length > 0),
+            () => Assert.Equal(1, viewOutput.Page),
+            () => Assert.Equal(viewOutput.Length, viewOutput.Entities.Length)
+
+        );
+    }
+
+    [Fact(DisplayName = "[Read]: Reads matched records")]
+    public async Task Read() {
+        Employee mock = Store(GenerateMock(RandomUtils.String(16)));
+        BatchOperationOutput<Employee, Employee> readOutput = await _service.Read(EntityBatchBehaviors.First, location => location.Id == mock.Id);
+
+        Assert.Multiple(
+            () => Assert.False(readOutput.Failed),
+            () => Assert.True(readOutput.Successes.First().Id > 0)
+        );
+    }
+}